requirements.txt
grpcio==1.57.0
grpcio-tools==1.57.0
protobuf>=4.21.6,<5.0.0
sounddevice
soundfile
numpy
scipy
